Updated to version 1.4.0.5

+++ MyGeotabAPIAdapter/DVIRLogManipulator.cs
@@ -21,6 +21,7 @@ namespace MyGeotabAPIAdapter
     /// </summary>
     class DVIRLogManipulator : BackgroundService
     {
+        const int ConnectivityRestorationCheckIntervalMilliseconds = 10000;
         const int DVIRDefectUpdateBatchSize = 5000;
         const int MyGeotabAPIAuthenticationCheckIntervalMilliseconds = 1000;
         readonly Logger logger = LogManager.GetCurrentClassLogger();
@@ -67,7 +68,7 @@ namespace MyGeotabAPIAdapter
             }
 
             // Add the remark.
-            DefectRemark defectRemark = new DefectRemark
+            DefectRemark defectRemark = new()
             {
                 User = new User { Id = Id.Create(dbDVIRDefectUpdate.RemarkUserId) },
                 DateTime = dbDVIRDefectUpdate.RemarkDateTime,
@@ -144,7 +145,7 @@ namespace MyGeotabAPIAdapter
                                 }
                                 else
                                 {
-                                    logger.Info($"No records retrieved from DVIRDefectUpdates table.");
+                                    logger.Debug($"No records retrieved from DVIRDefectUpdates table.");
                                     continue;
                                 }
 
@@ -174,18 +175,33 @@ namespace MyGeotabAPIAdapter
                                                     throw new ArgumentException($"No action to take because insufficent data was provided to either add a DefectRemark or update the RepairStatus of the subject DVIRDefect.");
                                                 }
 
+                                                var deleteCurrentDbDVIRDefectUpdateRecord = false;
                                                 try
                                                 {
                                                     // Update the DVIRLog in the MyGeotab database.
                                                     var result = await MyGeotabApiUtility.SetAsync<DVIRLog>(Globals.MyGeotabAPI, dvirLogToUpdate);
+                                                    deleteCurrentDbDVIRDefectUpdateRecord = true;
                                                     processedDbDVIRDefectUpdateCount += 1;
                                                 }
-                                                catch (Exception MyGeotabException)
+                                                catch (MyGeotabConnectionException)
                                                 {
-                                                    throw new Exception($"MyGeotab Exception encountered on Set<DVIRLog> call.", MyGeotabException);
+                                                    // Pass the MyGeotabConnectionException along so that the connectivity loss can be addressed and the subject DbDVIRDefectUpdate can be re-tried once connectivity is restored.
+                                                    throw;
                                                 }
-                                                // Delete the current DbDVIRDefectUpdate from the DVIRDefectUpdates table.
-                                                await DbDVIRDefectUpdateService.DeleteAsync(connectionInfo, dbDVIRDefectUpdate, commandTimeout);
+                                                catch (Exception exception)
+                                                {
+                                                    // Flag the current DbDVIRDefectUpdateRecord for deletion since this is not a connectivity-related exception and the same exception could occur repeatedly if the adapter is restarted and this record remains in the adapter database.
+                                                    deleteCurrentDbDVIRDefectUpdateRecord = true;
+                                                    throw new Exception($"MyGeotab Exception encountered on Set<DVIRLog> call.", exception);
+                                                }
+                                                finally
+                                                {
+                                                    if (deleteCurrentDbDVIRDefectUpdateRecord == true)
+                                                    {
+                                                        // Delete the current DbDVIRDefectUpdate from the DVIRDefectUpdates table.
+                                                        await DbDVIRDefectUpdateService.DeleteAsync(connectionInfo, dbDVIRDefectUpdate, commandTimeout);
+                                                    }
+                                                }
                                             }
                                         }
                                         else
@@ -210,7 +226,7 @@ namespace MyGeotabAPIAdapter
                                         }
                                         catch (Exception ex)
                                         {
-                                            throw new Exception($"Exception encountered while writing to {Globals.ConfigurationManager.DbFailedDVIRDefectUpdatesTableName} table or deleting from {Globals.ConfigurationManager.DbDVIRDefectUpdatesTableName} table.", ex);
+                                            throw new Exception($"Exception encountered while writing to {ConfigurationManager.DbFailedDVIRDefectUpdatesTableName} table or deleting from {ConfigurationManager.DbDVIRDefectUpdatesTableName} table.", ex);
                                         }
                                     }
                                 }
@@ -240,6 +256,10 @@ namespace MyGeotabAPIAdapter
                 catch (DatabaseConnectionException databaseConnectionException)
                 {
                     HandleException(databaseConnectionException, NLog.LogLevel.Error, $"{nameof(DVIRLogManipulator)} process caught an exception");
+                }
+                catch (MyGeotabConnectionException myGeotabConnectionException)
+                {
+                    HandleException(myGeotabConnectionException, NLog.LogLevel.Error, $"{nameof(DVIRLogManipulator)} process caught an exception");
                 }
                 catch (Exception ex)
                 {
@@ -272,8 +292,8 @@ namespace MyGeotabAPIAdapter
         }
 
         /// <summary>
-        /// Generates and logs an error message for the supplied <paramref name="exception"/>. Does NOT implement connectivity restoration logic like <see cref="Worker.WaitForConnectivityRestorationAsync(StateReason)"/> since the <see cref="Worker"/> will already be taking care of it.
-        /// /// </summary>
+        /// Generates and logs an error message for the supplied <paramref name="exception"/>. If the <paramref name="exception"/> is a <see cref="MyGeotabConnectionException"/> or a <see cref="DatabaseConnectionException"/>, the <see cref="WaitForConnectivityRestorationAsync(StateReason)"/> method will be executed after logging the error message.
+        /// </summary>
         /// <param name="exception">The <see cref="Exception"/>.</param>
         /// <param name="errorMessageLogLevel">The <see cref="NLog.LogLevel"/> to be used when logging the error message.</param>
         /// <param name="errorMessagePrefix">The start of the error message, which will be followed by the <see cref="Exception.Message"/>, <see cref="Exception.Source"/> and <see cref="Exception.StackTrace"/>.</param>
@@ -281,6 +301,15 @@ namespace MyGeotabAPIAdapter
         void HandleException(Exception exception, NLog.LogLevel errorMessageLogLevel, string errorMessagePrefix = "An exception was encountered")
         {
             Globals.LogException(exception, errorMessageLogLevel, errorMessagePrefix);
+
+            if (exception is MyGeotabConnectionException)
+            {
+                _ = WaitForConnectivityRestorationAsync(StateReason.MyGeotabNotAvailable);
+            }
+            if (exception is DatabaseConnectionException)
+            {
+                _ = WaitForConnectivityRestorationAsync(StateReason.DatabaseNotAvailable);
+            }
         }
 
         /// <summary>
@@ -409,6 +438,49 @@ namespace MyGeotabAPIAdapter
 
             logger.Trace($"End {methodBase.ReflectedType.Name}.{methodBase.Name}");
             return true;
+        }
+
+        /// <summary>
+        /// Iteratively tests MyGeotab or database connectivity, depending on which was lost, until connectivity is restored.
+        /// </summary>
+        /// <param name="reasonForConnectivityLoss">The reason for loss of connectivity.</param>
+        /// <returns></returns>        
+        async Task WaitForConnectivityRestorationAsync(StateReason reasonForConnectivityLoss)
+        {
+            MethodBase methodBase = MethodBase.GetCurrentMethod();
+            logger.Trace($"Begin {methodBase.ReflectedType.Name}.{methodBase.Name}");
+
+            StateMachine.CurrentState = State.Waiting;
+            StateMachine.Reason = reasonForConnectivityLoss;
+            logger.Warn($"******** CONNECTIVITY LOST. REASON: '{StateMachine.Reason}'. WAITING FOR RESTORATION OF CONNECTIVITY...");
+            while (StateMachine.CurrentState == State.Waiting)
+            {
+                // Wait for the prescribed interval between connectivity checks.
+                await Task.Delay(ConnectivityRestorationCheckIntervalMilliseconds);
+
+                logger.Warn($"{StateMachine.Reason}; continuing to wait for restoration of connectivity...");
+                switch (StateMachine.Reason)
+                {
+                    case StateReason.DatabaseNotAvailable:
+                        if (await StateMachine.IsDatabaseAccessibleAsync(connectionInfo) == true)
+                        {
+                            logger.Warn("******** CONNECTIVITY RESTORED.");
+                            StateMachine.CurrentState = State.Normal;
+                            continue;
+                        }
+                        break;
+                    case StateReason.MyGeotabNotAvailable:
+                        if (await StateMachine.IsMyGeotabAccessibleAsync() == true)
+                        {
+                            logger.Warn("******** CONNECTIVITY RESTORED.");
+                            StateMachine.CurrentState = State.Normal;
+                            continue;
+                        }
+                        break;
+                }
+            }
+
+            logger.Trace($"End {methodBase.ReflectedType.Name}.{methodBase.Name}");
         }
     }
 }
